Renamed the package "handlers" to "recipeHandlers"
