+++ src/main/java/craftedMods/lotr/nei/IntegratedRecipeHandlerFactory.java
@@ -4,7 +4,7 @@ import java.io.InputStream;
 import java.util.*;
 import java.util.function.Supplier;
 
-import craftedMods.lotr.nei.handlers.*;
+import craftedMods.lotr.nei.recipeHandlers.*;
 import craftedMods.recipes.api.*;
 import craftedMods.recipes.base.*;
 import lotr.client.gui.LOTRGuiCraftingTable;
