+++ src/main/java/craftedMods/lotr/nei/recipeHandlers/AlloyForgeRecipeHandler.java
@@ -1,8 +1,8 @@
-package craftedMods.lotr.nei.handlers;
+package craftedMods.lotr.nei.recipeHandlers;
 
 import java.util.*;
 
-import craftedMods.lotr.nei.handlers.AlloyForgeRecipeHandler.AlloyForgeRecipe;
+import craftedMods.lotr.nei.recipeHandlers.AlloyForgeRecipeHandler.AlloyForgeRecipe;
 import craftedMods.recipes.api.*;
 import craftedMods.recipes.api.utils.RecipeHandlerRendererUtils;
 import craftedMods.recipes.api.utils.RecipeHandlerRendererUtils.EnumProgressBarDirection;
