+++ src/main/java/craftedMods/lotr/nei/recipeHandlers/KebabRecipeHandler.java
@@ -1,4 +1,4 @@
-package craftedMods.lotr.nei.handlers;
+package craftedMods.lotr.nei.recipeHandlers;
 
 import java.util.*;
