+++ src/main/java/craftedMods/lotr/nei/recipeHandlers/UnsmelteryRecipeHandler.java
@@ -1,4 +1,4 @@
-package craftedMods.lotr.nei.handlers;
+package craftedMods.lotr.nei.recipeHandlers;
 
 import java.lang.reflect.Method;
 import java.util.*;
@@ -6,7 +6,7 @@ import java.util.*;
 import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 import cpw.mods.fml.common.gameevent.TickEvent;
-import craftedMods.lotr.nei.handlers.UnsmelteryRecipeHandler.UnsmelteryRecipe;
+import craftedMods.lotr.nei.recipeHandlers.UnsmelteryRecipeHandler.UnsmelteryRecipe;
 import craftedMods.recipes.NEIExtensions;
 import craftedMods.recipes.api.*;
 import craftedMods.recipes.api.utils.*;
